requirements.txt
streamlit
google-genai
pandas
openpyxl
thefuzz
python-levenshtein
pillow